showing daily weather from api

+++ app/src/main/java/com/almaz/weather_aa/core/model/DailyWeather.kt
@@ -5,15 +5,15 @@ import com.google.gson.annotations.SerializedName
 
 data class DailyWeather(
     @SerializedName("app_max_temp")
-    val appMaxTemp: Int,
+    val appMaxTemp: Double,
     @SerializedName("app_min_temp")
-    val appMinTemp: Int,
+    val appMinTemp: Double,
     @SerializedName("clouds")
-    val clouds: Int,
+    val clouds: Double,
     @SerializedName("datetime")
     val datetime: String,
     @SerializedName("dewpt")
-    val dewpt: Int,
+    val dewpt: Double,
     @SerializedName("max_dhi")
     val maxDhi: String,
     @SerializedName("max_temp")
@@ -23,31 +23,31 @@ data class DailyWeather(
     @SerializedName("moon_phase")
     val moonPhase: Double,
     @SerializedName("moonrise_ts")
-    val moonriseTs: Int,
+    val moonriseTs: Double,
     @SerializedName("moonset_ts")
-    val moonsetTs: Int,
+    val moonsetTs: Double,
     @SerializedName("pod")
     val pod: String,
     @SerializedName("pop")
-    val pop: Int,
+    val pop: Double,
     @SerializedName("precip")
     val precip: Double,
     @SerializedName("pres")
-    val pres: Int,
+    val pres: Double,
     @SerializedName("rh")
-    val rh: Int,
+    val rh: Double,
     @SerializedName("slp")
     val slp: Double,
     @SerializedName("snow")
     val snow: Double,
     @SerializedName("snow_depth")
-    val snowDepth: Int,
+    val snowDepth: Double,
     @SerializedName("sunrise_ts")
-    val sunriseTs: Int,
+    val sunriseTs: Double,
     @SerializedName("sunset_ts")
-    val sunsetTs: Int,
+    val sunsetTs: Double,
     @SerializedName("temp")
-    val temp: Int,
+    val temp: Double,
     @SerializedName("timestamp_local")
     val timestampLocal: String,
     @SerializedName("timestamp_utc")
@@ -57,7 +57,7 @@ data class DailyWeather(
     @SerializedName("uv")
     val uv: Double,
     @SerializedName("vis")
-    val vis: Int,
+    val vis: Double,
     @SerializedName("weather")
     val weather: WeatherDetails,
     @SerializedName("wind_cdir")
@@ -65,7 +65,7 @@ data class DailyWeather(
     @SerializedName("wind_cdir_full")
     val windCdirFull: String,
     @SerializedName("wind_dir")
-    val windDir: Int,
+    val windDir: Double,
     @SerializedName("wind_spd")
     val windSpd: Double
 )
